added in copy_usr_bin directory to ease deploying files to templates
sudo qubesctl --skip-dom0 --targets=<DESTVM> state.apply usr-bin-copy saltenv=user

easily deploy executable scripts to VMs /usr/bin

included is rolo: the Read-Only LOopback deployer
use this in a domU with the parameter of a path and an exfat|xfs block
device will be created with that path's contents and set up
as a loopback device. qubes will detect this and show a notification as
well as allow it to be attached to a VM of your choice.

exfat (default) is a good choice if moving files to windows
xfs is a godo choice when staying among *nix

FS_TYPE=exfat rolo /home/user/Public
FS_TYPE=xfs rolo /home/user/Public